+++ report-front-end/src/common/helpers/store.ts
@@ -1,6 +1,6 @@
 import { createStore } from "redux";
+import { DEFAULT_QUERY_CONFIG, LOCALSTORAGE_KEY } from "../constant/constants";
 import { ActionType, UserAction, UserState } from "./types";
-import { DEFAULT_QUERY_CONFIG, LOCALSTORAGE_KEY } from "../../common/constant/constants";
 
 const defaultUserState: UserState = {
   userId: "",
