--- report-front-end/src/components/config-panel/types.ts
@@ -1,15 +1,0 @@
-export enum ActionType {
-  Update = "Update",
-  Delete = "Delete",
-  UpdateConfig = "UpdateConfig",
-}
-
-export type UserState = {
-  userId: string;
-  displayName: string;
-  email: string;
-  loginExpiration: number;
-  queryConfig: any;
-};
-
-export type UserAction = { type: ActionType; state?: any };
