+++ report-front-end/src/main.tsx
@@ -4,7 +4,7 @@ import ReactDOM from "react-dom/client";
 import { Storage } from "./common/helpers/storage";
 import "@cloudscape-design/global-styles/index.css";
 import { Provider } from "react-redux";
-import userReduxStore from "./components/config-panel/store";
+import userReduxStore from "./common/helpers/store";
 import AppConfigured from "./pages/login-page";
 
 const root = ReactDOM.createRoot(
